made requested changes from aman
